BE/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Environment and configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# MongoDB driver
motor==3.3.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production dependencies
gunicorn==21.2.0
